Move azul.plugins.metadata.hca.contributor_matrices into ...service
